Added separated files for word, excel, pp tests.
